Add conversion code to the BOP format (#12)
Summary:
This PR contains 3 scripts to convert the HOT3D-Clips dataset from its native format to the BOP format. The BOP format had to be extended because the original format only included RGB and depth modalities. This extension made it possible to include a varying number of modalities (RGB, gray1, gray2) with separate annotations for each modality. The dataset has already been converted using these scripts and uploaded to the [BOP website](https://bop.felk.cvut.cz/datasets/). These scripts went through several revisions, as the conversion process required multiple changes to adapt the dataset to the new standard format, which was initially not strictly defined and required frequent adjustments. The scripts in this PR are isolated and do not make any changes to the existing code in the repository.

The 3 Scripts are:
1. hot3d_models_to_bop.py : Script that converts the 3D object (full) models from the GLTF format to the PLY format
2. hot3d_models_eval_to_bop.py: Script to convert the eval object model
3. hot3d_clips_to_bop_scenewise.py: Script to convert the dataset itself

Pull Request resolved: https://github.com/facebookresearch/hot3d/pull/12

Reviewed By: prithvirb

Differential Revision: D62526222

Pulled By: SeaOtocinclus

fbshipit-source-id: 8ff1cb1f0368d91d7b1ad63300ecf7e9146cdb5d